refactor: db/index only exports
